requirements.txt
opencv-contrib-python
Pillow
numpy
pandas
mysql-connector-python
openpyxl
